requirements.txt
beautifulsoup4
requests
transformers
torch
groq
python-telegram-bot
discord-webhook
reportlab
tenacity
PyYAML
faiss-cpu
accelerate
datasets